+++ trooptrak_final_application/android/gradle.properties
@@ -1,3 +1,4 @@
 org.gradle.java.home=C:\\Program Files\\Java\\jdk-18.0.2.1
+org.gradle.jvmargs=-Xmx4096m -Dfile.encoding=UTF-8
 android.useAndroidX=true
 android.enableJetifier=true
