+++ trooptrak_final_application/lib/counter_model.dart
@@ -1,0 +1,12 @@
+import 'package:flutter/material.dart';
+
+class CounterModel extends ChangeNotifier {
+  int _counter = 0;
+
+  int get counter => _counter;
+
+  void incrementCounter() {
+    _counter++;
+    notifyListeners();
+  }
+}
